complete the taxes and the transfer money ehn customer accept a seller

+++ JobDone/Models/SellerAcceptRequest/SellerAcceptRequestImp.cs
@@ -4,6 +4,8 @@ using JobDone.Models.Category;
 using JobDone.Models.Customer;
 using JobDone.Models.Order;
 using JobDone.Models.OrderByCustomer;
+using JobDone.Models.Seller;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using NuGet.Packaging;
@@ -15,11 +17,15 @@ namespace JobDone.Models.SellerAcceptRequest
     {
         private readonly JobDoneContext _db;
         private readonly DbSet<SellerAcceptRequestModel> _sellersWhoAccept;
+        private readonly DbSet<SellerModel> _sellers;
+        private readonly DbSet<AdminWalletModel> _adminWallet;
 
         public SellerAcceptRequestImp(JobDoneContext db)
         {
             _db = db;
             _sellersWhoAccept = _db.SellerAcceptRequestModels;
+            _sellers = _db.SellerModels;
+            _adminWallet = _db.AdminWalletModels;
 
         }
 
@@ -39,36 +45,64 @@ namespace JobDone.Models.SellerAcceptRequest
             return sellers;
         }
 
-        public async Task<SellerAcceptRequestModel> RemoveSeller(int sellerId, int OrderByCustomerId)
+        public async Task<SellerAcceptRequestModel>? RemoveSeller(int sellerId, int OrderByCustomerId)
         {
-            SellerAcceptRequestModel request = _db.SellerAcceptRequestModels.FirstOrDefault(
+            SellerAcceptRequestModel? request = _db.SellerAcceptRequestModels.FirstOrDefault(
                 SAM => SAM.SellerIdFk == sellerId && SAM.OrderByCustomerIdFk == OrderByCustomerId);
-            _db.SellerAcceptRequestModels.Remove(request);
-            _db.SaveChanges();
+            try
+            {
+                if (request != null)
+                {
+                    _db.SellerAcceptRequestModels.Remove(request);
+                    _db.SaveChanges();
+                    return request;
+                }
+                return null;
+            }
+            catch
+            {
+                return null;
+            }
 
-            return request;
         }
         
         public async Task<bool> AcceptSeller(int sellerId, int OrderByCustomerId)
         {
+            //website taxes percentage
+            decimal taxPercentage = 5m / 100m;
+            //get the sellers from specific request who accepted the order except the seller that the customer accepts
+
             List<SellerAcceptRequestModel> requests = _db.SellerAcceptRequestModels.Where(
                 SAM => SAM.SellerIdFk != sellerId && SAM.OrderByCustomerIdFk == OrderByCustomerId).ToList();
 
-                OrderByCustomerModel orderByCustomer = _db.OrderByCustomerModels.FirstOrDefault(o => o.Id == OrderByCustomerId);
+            OrderByCustomerModel? orderByCustomer = _db.OrderByCustomerModels.FirstOrDefault(o => o.Id == OrderByCustomerId);
 
-            CustomerModel customer = _db.CustomerModels.FirstOrDefault(c => c.Id == orderByCustomer.CustomerIdFk);
+            CustomerModel? customer = _db.CustomerModels.FirstOrDefault(c => c.Id == orderByCustomer.CustomerIdFk);
+
+            SellerModel? seller = await _sellers.FirstOrDefaultAsync(s => s.Id == sellerId);
+
+            AdminWalletModel adminWallet = await _adminWallet.FirstAsync();
             if (customer.Wallet >= orderByCustomer.Price)
             {
+                // take the order price from the customer
                  customer.Wallet -= orderByCustomer.Price;
-                 _db.CustomerModels.Update(customer);
+
+                // transfer the money of the order to the seller and take 5% for the website
+                 seller.Wallet += (orderByCustomer.Price - (taxPercentage * orderByCustomer.Price));
+
+                // transfer 5% of the order money to the admin wallet
+                adminWallet.Balance += (taxPercentage * orderByCustomer.Price);
+                _db.CustomerModels.Update(customer);
+                _adminWallet.Update(adminWallet);
+                _sellers.Update(seller);
                 try
                 {
-                
+                    //remove the seller that the customer accepts from the request list
                     SellerAcceptRequestModel request = RemoveSeller(sellerId, OrderByCustomerId).Result;
 
 
-
-                    CategoryModel category = _db.CategoryModels.FirstOrDefault(c => c.Id == orderByCustomer.CategoryIdKf);
+                    //get the category related to the request
+                    CategoryModel? category = _db.CategoryModels.FirstOrDefault(c => c.Id == orderByCustomer.CategoryIdKf);
 
 
                 OrderModel orderModel = new OrderModel()
